* Few fixes

+++ persistence/src/main/java/com/redshape/persistence/entities/IEntity.java
@@ -3,15 +3,12 @@ package com.redshape.persistence.entities;
 import java.io.Serializable;
 
 /**
- * Интерфейс для представления абстрактного объекта из хранилища постоянных
- * объектов
- *
  * @author nikelin
  */
 public interface IEntity extends Serializable {
 
-    public Integer getId();
+    public Long getId();
 
-    public void setId( Integer id );
-
+    public void setId( Long id );
+    
 }
